third_party/blink/web_tests/fast/events/event-on-created-document-expected.txt

PASS window.caughtFoo is ['foo@target', 'foo@root']
PASS window.caughtFoo is []
PASS window.caughtFoo is ['foo@target', 'foo@root']
PASS window.caughtFoo is []
PASS window.caughtFoo is ['foo@target', 'foo@root']
PASS window.caughtFoo is []
PASS window.caughtFoo is ['foo@target', 'foo@root']
PASS window.caughtFoo is []
PASS successfullyParsed is true

TEST COMPLETE
